test/requirements.txt
paho-mqtt==1.6.1
msgpack==1.0.7